refactor(nvim): split keymaps; add NERDTree and Easymotion; move ALE maps
Replace keymaps/core with editor + integrations; register ALE jumps and
haskell <leader>? in plugins/ale.lua; add preservim/nerdtree and
vim-easymotion so F6 and easymotion binds match installed plugins.
