+++ lazy-lock.json
@@ -24,11 +24,12 @@
   "hexmode": { "branch": "master", "commit": "550cae65fdbd06d61d757dc6fe5a9be5be2e3ef4" },
   "html5.vim": { "branch": "master", "commit": "485f2cd62162c81e56d8604b4c630f0b5ef69d1f" },
   "jspc.vim": { "branch": "master", "commit": "a957665464020b2b4003f2a31d941ef3b95ffa40" },
-  "lazy.nvim": { "branch": "main", "commit": "85c7ff3711b730b4030d03144f6db6375044ae82" },
+  "lazy.nvim": { "branch": "main", "commit": "306a05526ada86a7b30af95c5cc81ffba93fef97" },
   "mango.vim": { "branch": "master", "commit": "7af1e00e9068d4c1010a05626030eb0bf64365d7" },
   "markdown-preview.nvim": { "branch": "master", "commit": "a923f5fc5ba36a3b17e289dc35dc17f66d0548ee" },
   "mustang-vim": { "branch": "master", "commit": "6533d7d21bf27cae94d9c2caa575f627f003dfd5" },
   "neco-ghc": { "branch": "master", "commit": "699897c2f4ba82c4fd2be6b93c9a2e8e548efe4e" },
+  "nerdtree": { "branch": "master", "commit": "690d061b591525890f1471c6675bcb5bdc8cdff9" },
   "nova.vim": { "branch": "master", "commit": "48fdce03b839bc93dbae2dda996d5533dd1d53ff" },
   "nvim-cmp": { "branch": "main", "commit": "a1d504892f2bc56c2e79b65c6faded2fd21f3eca" },
   "nvim-treesitter": { "branch": "main", "commit": "6878ae017d691a0df1c49d8a091c88a33e4c2d50" },
@@ -49,6 +50,7 @@
   "vim-composer": { "branch": "master", "commit": "7ad79699a52f9974d8cc133d4a6c9bf988bb578c" },
   "vim-dispatch": { "branch": "master", "commit": "a2ff28abdb2d89725192db5b8562977d392a4d3f" },
   "vim-easy-align": { "branch": "master", "commit": "9815a55dbcd817784458df7a18acacc6f82b1241" },
+  "vim-easymotion": { "branch": "master", "commit": "b3cfab2a6302b3b39f53d9fd2cd997e1127d7878" },
   "vim-erlang-runtime": { "branch": "master", "commit": "976b10b74c75d1297bef5b2a0e56ff2f75343ee8" },
   "vim-fugitive": { "branch": "master", "commit": "3b753cf8c6a4dcde6edee8827d464ba9b8c4a6f0" },
   "vim-gitgutter": { "branch": "main", "commit": "55b368d4c68541775b4b253d3c98d3145b695c24" },
