+++ lua/keymaps.lua
@@ -4,7 +4,8 @@ function M.setup()
 	-- Must run before lazy.nvim (see require order in init.vim).
 	vim.g.mapleader = " "
 
-	require("keymaps.core").setup()
+	require("keymaps.editor").setup()
+	require("keymaps.integrations").setup()
 	require("keymaps.coc").setup()
 end
