+++ settings/mappings.vim
@@ -1,2 +1,2 @@
-" Mappings live in Lua: lua/keymaps.lua → keymaps/core.lua, keymaps/coc.lua
+" Mappings live in Lua: lua/keymaps.lua → keymaps/editor.lua, integrations.lua, coc.lua
 " g:mapleader is set in lua/keymaps.lua (before lazy.nvim).
